Refresh against documentation
Most of this is just new features and documentation updates

There is a smallish break in the IdentityAPI caused by an error added to
the marketo documentation. When removing a legacy argument, they seem to
have also broken the "required" documentation. Both the removal and the
conversion of the parameters could be breaking if you where extending
the IdentyAPI so take a look.

Also, new phpstan changes caught a new bit of overly generalized code.
Excluded for the moment.